Move CoreDataFeedStore implementation of the FeedStore protocol t…
…o separate file
